requirements.txt
amqp==5.0.6
aniso8601==8.1.1
asgiref==3.4.1
backports.entry-points-selectable==1.1.1
billiard==3.6.4.0
celery==5.1.2
certifi==2021.10.8
cfgv==3.3.1
charset-normalizer==2.0.7
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
distlib==0.3.4
Django==3.2.6
django-cors-headers==3.10.0
django-environ==0.8.1
django-filter==2.4.0
django-qr-code==2.2.0
djangorestframework==3.12.4
djangorestframework-camel-case==1.2.0
djangorestframework-simplejwt==4.7.2
filelock==3.4.0
graphene==3.0b7
graphene-django==3.0.0b7
graphql-core==3.1.5
graphql-relay==3.1.0
gunicorn==20.1.0
identify==2.4.0
idna==3.3
kombu==5.1.0
nodeenv==1.6.0
Pillow==8.3.1
platformdirs==2.4.0
pre-commit==2.16.0
promise==2.3
prompt-toolkit==3.0.20
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.1
PyYAML==6.0
redis==3.5.3
requests==2.26.0
segno==1.3.3
six==1.16.0
sqlparse==0.4.1
stripe==2.61.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.7
vine==5.0.0
virtualenv==20.10.0
wcwidth==0.2.5
